pyproject.toml
[project]
name = "langgraph-tutorials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotty-dictionary>=1.2.0",
    "ipykernel>=6.30.1",
    "langchain-core>=0.3.75",
    "langchain-openai>=0.3.32",
    "langchain-tavily>=0.2.11",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.6",
    "langgraph-supervisor>=0.0.29",
    "langsmith>=0.4.19",
    "loguru>=0.7.3",
    "openai>=1.102.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.0",
]
